change the folders names
