+++ appium-examples/src/test/java/com/emusim/biometric_login/AndroidSettings.java
@@ -1,4 +1,4 @@
-package com.native_app.biometric_login;
+package com.emusim.biometric_login;
 
 import io.appium.java_client.MobileBy;
 import io.appium.java_client.android.AndroidDriver;
