+++ appium-examples/src/test/java/com/realdevice/unifiedplatform/SauceStorage.sh
