+++ appium-examples/src/test/java/com/realdevice/unifiedplatform/image_injection/ImageInjectionAndroidTest.java
@@ -1,37 +1,32 @@
-package com.native_app.image_injection;
+package com.realdevice.unifiedplatform.image_injection;
 
-import io.appium.java_client.ios.IOSDriver;
+import io.appium.java_client.MobileBy;
+import io.appium.java_client.android.AndroidDriver;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.api.extension.TestWatcher;
-import org.openqa.selenium.By;
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.MutableCapabilities;
-import org.openqa.selenium.WebElement;
+import org.openqa.selenium.*;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-
 import java.io.IOException;
-import java.lang.reflect.Method;
-import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.concurrent.TimeUnit;
 
 import static helpers.Constants.region;
 
-public class ImageInjectionIosTest {
+public class ImageInjectionAndroidTest {
 
-    By usernameID = By.id("test-Username");
-    By passwordID = By.id("test-Password");
-    By submitButton = By.id("test-LOGIN");
-    String ProductTitleSelector = "type==\"XCUIElementTypeStaticText\" && name==\"PRODUCTS\"";
-    By testMenu = By.name("test-Menu");
-    By testMenuItemWebView = By.name("test-WEBVIEW");
-    By testMenuItemQRCode = By.name("test-QR CODE SCANNER");
-    By acceptCameraButton = By.name("OK");
+    String usernameID = "test-Username";
+    String passwordID = "test-Password";
+    String submitButtonID = "test-LOGIN";
 
-    protected IOSDriver driver;
+    By ProductTitle = By.xpath("//android.widget.TextView[@text='PRODUCTS']");
+
+    By testMenu = By.xpath("//android.view.ViewGroup[@content-desc='test-Menu']");
+    By testMenuItemQRCode = By.xpath("//android.view.ViewGroup[@content-desc='test-QR CODE SCANNER']");
+    By testMenuItemWebView = By.xpath("//android.view.ViewGroup[@content-desc='test-WEBVIEW']");
+
+    protected AndroidDriver driver;
 
     @RegisterExtension
     public MyTestWatcher myTestWatcher = new MyTestWatcher();
@@ -51,33 +46,31 @@ public class ImageInjectionIosTest {
         } else {
             sauceUrl = "@ondemand.us-west-1.saucelabs.com:443";
         }
-
         String SAUCE_REMOTE_URL = "https://" + username + ":" + accesskey + sauceUrl +"/wd/hub";
         URL url = new URL(SAUCE_REMOTE_URL);
 
-
         MutableCapabilities capabilities = new MutableCapabilities();
-        capabilities.setCapability("deviceName", "iPhone 8*");
-        capabilities.setCapability("platformName", "iOS");
-        capabilities.setCapability("automationName", "XCUITEST");
+        capabilities.setCapability("deviceName", "Samsung Galaxy S10");
+
+        capabilities.setCapability("platformName", "Android");
+        capabilities.setCapability("automationName", "UiAutomator2");
         capabilities.setCapability("name", methodName);
-        //      You can use  storage:filename=" +appName if you uploaded your app to Saucd Storage
+//      You can use  storage:filename=" +appName if you uploaded your app to Saucd Storage
 //        capabilities.setCapability("app", "storage:filename=" +appName);
-        capabilities.setCapability("app", "https://github.com/saucelabs/sample-app-mobile/releases/download/2.7.1/iOS.RealDevice.SauceLabs.Mobile.Sample.app.2.7.1.ipa");
+        capabilities.setCapability("app", "https://github.com/saucelabs/sample-app-mobile/releases/download/2.7.1/Android.SauceLabs.Mobile.Sample.app.2.7.1.apk");
+        capabilities.setCapability("appWaitActivity", "com.swaglabsmobileapp.MainActivity");
 
         capabilities.setCapability("sauceLabsImageInjectionEnabled", true);
-        capabilities.setCapability("autoAcceptAlerts", true);
+        capabilities.setCapability("autoGrantPermissions", true);
 
         // Launch remote browser and set it as the current thread
-        driver = new IOSDriver(url, capabilities);
-        driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
-
+        driver = new AndroidDriver(url, capabilities);
     }
 
     @Test
     @DisplayName("imageInjectionScanQRcode")
     @Tag("imageInjection")
-    @Tag("ios")
+    @Tag("Android")
     public void imageInjectionScanQRcode() throws InterruptedException {
         System.out.println("Sauce - start test imageInjection_scan_QR_code");
 
@@ -86,20 +79,18 @@ public class ImageInjectionIosTest {
 
         // Verificsation
         Assertions.assertTrue(isOnProductsPage());
+
         // Select QR Code Scanner from the menu
         clickMenu();
         selecMenuQRCodeScanner();
 
-        // Accept access if asked
-        acceptCameraAccess();
-
-        Utils utils = new Utils();
         // inject the image - provide the transformed image to the device with this command
+        Utils utils = new Utils();
         String qrCodeImage = utils.encoder("src/test/java/com/native_app/image_injection/images/qr-code.png");
         ((JavascriptExecutor)driver).executeScript("sauce:inject-image=" + qrCodeImage);
 
         // Verify that the browser is running
-        utils.isIosApplicationRunning(driver, "com.apple.mobilesafari");
+        utils.isAndroidBrowserOpened(driver);
 
         // This is not need only for the video
         try
@@ -110,6 +101,7 @@ public class ImageInjectionIosTest {
         {
             Thread.currentThread().interrupt();
         }
+
     }
 
     public class MyTestWatcher implements TestWatcher {
@@ -120,7 +112,7 @@ public class ImageInjectionIosTest {
                 ((JavascriptExecutor) driver).executeScript("sauce:job-result=passed");
             } catch (Exception ignored) {
             } finally {
-                driver.quit();
+                    driver.quit();
             }
         }
 
@@ -136,19 +128,23 @@ public class ImageInjectionIosTest {
         }
     }
 
-
     private void login(String user, String pass) {
         try {
+            driver.context("NATIVE_APP");
 
-            WebElement usernameEdit = driver.findElement(usernameID);
+            WebDriverWait wait = new WebDriverWait(driver, 10);
+            final WebElement usernameEdit = wait.until(ExpectedConditions.visibilityOfElementLocated(new MobileBy.ByAccessibilityId(usernameID)));
+
             usernameEdit.click();
             usernameEdit.sendKeys(user);
 
-            WebElement passwordEdit = driver.findElement(passwordID);
+
+            WebElement passwordEdit = (WebElement) driver.findElementByAccessibilityId(passwordID);
             passwordEdit.click();
             passwordEdit.sendKeys(pass);
 
-            driver.findElement(submitButton).click();
+            WebElement submitButton = (WebElement) driver.findElementByAccessibilityId(submitButtonID);
+            submitButton.click();
 
         } catch (Exception e) {
             System.out.println("*** Problem to login: " + e.getMessage());
@@ -156,31 +152,26 @@ public class ImageInjectionIosTest {
     }
 
     private boolean isOnProductsPage() {
-        return driver.findElementByIosNsPredicate(ProductTitleSelector).isDisplayed();
+
+        //Create an instance of a Selenium explicit wait so that we can dynamically wait for an element
+        WebDriverWait wait = new WebDriverWait(driver, 5);
+
+        //wait for the product field to be visible and store that element into a variable
+        try {
+            wait.until(ExpectedConditions.visibilityOfElementLocated(ProductTitle));
+        } catch (TimeoutException e){
+            return false;
+        }
+        return true;
     }
 
     private void clickMenu() {
         driver.findElement(testMenu).click();
     }
-
     private void selecMenuQRCodeScanner() {
         WebDriverWait wait = new WebDriverWait(driver, 5);
 
         WebElement QCCodeMenu = wait.until(ExpectedConditions.visibilityOfElementLocated(testMenuItemQRCode));
         QCCodeMenu.click();
     }
-
-    public void acceptCameraAccess(){
-        // In Sauce Labs (Legacy) RDC Android permissions are enabled by default.
-        /// iOS needs to be done with automation
-        try {
-            WebDriverWait wait = new WebDriverWait(driver, 3);
-            WebElement acceptCameraBtn = wait.until(ExpectedConditions.visibilityOfElementLocated(acceptCameraButton));
-            acceptCameraBtn.click();
-        } catch (Exception e) {
-            // Do nothing, the alert was not shown
-        }
-    }
-
-
 }
