+++ appium-examples/src/test/java/com/realdevice/unifiedplatform/image_injection/README.md
@@ -20,10 +20,6 @@ Be aware of the fact that you need the build for the iOS real device. So please 
 
 > The advice is to download the files to an `apps` folder in the root of this folder.
 
-### Upload apps to Sauce Storage
-* If you want to use Native apps you need to upload the apps to the Sauce Storage.
-For more information on this step please visit: [Application Storage](https://wiki.saucelabs.com/display/DOCS/Application+Storage).
-* In the app capability you can use storage:filename=<file-name>. For more information on this step please visit: [Using Application Storage with Automated Test Builds](https://wiki.saucelabs.com/display/DOCS/Application+Storage#ApplicationStorage-UsingApplicationStoragewithAutomatedTestBuilds) section of [Application Storage](https://wiki.saucelabs.com/display/DOCS/Application+Storage)
 ## Run tests on Sauce Labs Android real devices in the New Sauce Labs UI
 If you want to run the tests on Sauce Labs real devices in the **New Sauce Labs UI** then you can run the Android test with
