+++ appium-examples/src/test/java/com/realdevice/unifiedplatform/image_injection/images/qr-code.png
